+++ packages/fsapp/src/beta-app/index.ts
@@ -3,3 +3,5 @@ export * from './modal';
 export * from './router';
 export * from './shell.web';
 export * from './store';
+
+export * from './lib/use-dependency';