+++ packages/fscomponents/src/components/MultiCarousel/MultiCarousel.web.tsx
@@ -89,6 +89,10 @@ const styles = StyleSheet.create({
 // tslint:disable-next-line: cyclomatic-complexity
 export const MultiCarousel = <ItemT, >(props: MultiCarouselProps<ItemT>) => {
   const {
+    accessible,
+    accessibilityHint,
+    accessibilityLabel,
+    accessibilityRole,
     data,
     renderItem,
     PageIndicatorComponent = DEFAULT_PAGE_INDICATOR_COMPONENT,
@@ -256,6 +260,11 @@ export const MultiCarousel = <ItemT, >(props: MultiCarouselProps<ItemT>) => {
             setAnimating(false);
             console.warn('animatedScrollTo error', e);
           });
+        // Hacky fix for a hacky Safari release. This fixes a scroll issue
+        // on iOS 14.
+        setTimeout(() => {
+          scrollViewElement.scrollLeft = nextIndex * pageWidth;
+        }, 0);
       }
     },
     [animating, currentIndex, pageWidth, setAnimating, onSlideChange]
@@ -445,6 +454,10 @@ export const MultiCarousel = <ItemT, >(props: MultiCarouselProps<ItemT>) => {
         onScroll={handleScroll}
       >
         <View
+          accessible={accessible}
+          accessibilityHint={accessibilityHint}
+          accessibilityLabel={accessibilityLabel}
+          accessibilityRole={accessibilityRole}
           style={{
             width: centerMode ? calculatedPeekSize : 0
           }}