+++ packages/fsweb/webpack.config.js
@@ -9,6 +9,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
 const BabelPluginTransformImports = require('babel-plugin-transform-imports');
 const BabelPluginReactNativeWeb = require('babel-plugin-react-native-web');
+const BabelPluginProposalPrivatePropertyInObject = require('@babel/plugin-proposal-private-property-in-object');
 const BabelPluginProposalClassProperties = require('@babel/plugin-proposal-class-properties');
 const BabelPluginProposalPrivateMethods = require('@babel/plugin-proposal-private-methods');
 const escapedSep = '\\' + path.sep;
@@ -105,6 +106,7 @@ const globalConfig = {
                     [BabelPluginReactNativeWeb, {
                       commonjs: false
                     }],
+                    [BabelPluginProposalPrivatePropertyInObject, { loose: true }],
                     [BabelPluginProposalClassProperties, { loose: true }],
                     [BabelPluginProposalPrivateMethods, { loose: true }]
                   ]
@@ -150,6 +152,7 @@ const globalConfig = {
                     [BabelPluginReactNativeWeb, {
                       commonjs: false
                     }],
+                    [BabelPluginProposalPrivatePropertyInObject, { loose: true }],
                     [BabelPluginProposalClassProperties, { loose: true }],
                     [BabelPluginProposalPrivateMethods, { loose: true }]
                   ]
@@ -173,28 +176,30 @@ const globalConfig = {
                 {
                   loader: require.resolve('postcss-loader'),
                   options: {
-                    // Necessary for external CSS imports to work
-                    // https://github.com/facebookincubator/create-react-app/issues/2677
-                    ident: 'postcss',
-                    plugins: () => [
-                      require('postcss-flexbugs-fixes'),
-                      autoprefixer({
-                        browsers: [
-                          '>1%',
-                          'last 4 versions',
-                          'Firefox ESR',
-                          'not ie < 9' // React doesn't support IE8 anyway
-                        ],
-                        flexbox: 'no-2009'
-                      }),
-                      require('cssnano')({
-                        preset: ['default', {
-                          discardComments: {
-                            removeAll: true,
-                          },
-                        }]
-                      })
-                    ]
+                    postcssOptions: {
+                      // Necessary for external CSS imports to work
+                      // https://github.com/facebookincubator/create-react-app/issues/2677
+                      ident: 'postcss',
+                      plugins: () => [
+                        require('postcss-flexbugs-fixes'),
+                        autoprefixer({
+                          browsers: [
+                            '>1%',
+                            'last 4 versions',
+                            'Firefox ESR',
+                            'not ie < 9' // React doesn't support IE8 anyway
+                          ],
+                          flexbox: 'no-2009'
+                        }),
+                        require('cssnano')({
+                          preset: ['default', {
+                            discardComments: {
+                              removeAll: true,
+                            },
+                          }]
+                        })
+                      ]
+                    }
                   }
                 }
               ]